requirements.txt
flwr>=1.5.0
numpy
Pillow
protobuf
grpcio